Update dependencies and version
The dependencies for rake and rspec have been updated to more recent versions to ensure compatibility with new behavior changes in these libraries. Additionally, the bundler dependency has been removed as it was no longer necessary for the development environment. The minor version of the repost gem has been incremented to reflect these changes.